+++ programs/cp-swap/src/curve/fees.rs
@@ -1,6 +1,6 @@
 //! All fee information, to be used for validation currently
 
-pub const FEE_RATE_DENOMINATOR_VALUE: u32 = 1_000_000;
+pub const FEE_RATE_DENOMINATOR_VALUE: u64 = 1_000_000;
 
 pub struct Fees {}
 
@@ -24,7 +24,7 @@ pub fn floor_div(token_amount: u128, fee_numerator: u128, fee_denominator: u128)
 
 impl Fees {
     /// Calculate the trading fee in trading tokens
-    pub fn trading_fee(amount: u128, trade_fee_rate: u32) -> Option<u128> {
+    pub fn trading_fee(amount: u128, trade_fee_rate: u64) -> Option<u128> {
         ceil_div(
             amount,
             u128::from(trade_fee_rate),
@@ -33,7 +33,7 @@ impl Fees {
     }
 
     /// Calculate the owner trading fee in trading tokens
-    pub fn protocol_fee(amount: u128, protocol_fee_rate: u32) -> Option<u128> {
+    pub fn protocol_fee(amount: u128, protocol_fee_rate: u64) -> Option<u128> {
         floor_div(
             amount,
             u128::from(protocol_fee_rate),
@@ -42,7 +42,7 @@ impl Fees {
     }
 
     /// Calculate the owner trading fee in trading tokens
-    pub fn fund_fee(amount: u128, fund_fee_rate: u32) -> Option<u128> {
+    pub fn fund_fee(amount: u128, fund_fee_rate: u64) -> Option<u128> {
         floor_div(
             amount,
             u128::from(fund_fee_rate),
